Chore: API 디렉토리 posts에서 post로 이동
