+++ src/app/api/v1/posts/route.js
@@ -52,36 +52,4 @@ async function GET(request) {
   }
 }
 
-/**
- * 포스트 생성 API
- * @URL /api/v1/posts
- * @param request
- */
-async function POST(request) {
-  await dbConnect();
-
-  try {
-    const { title, content, author } = await request.json();
-
-    if (!title || !content || !author) {
-      throw createError(
-        ERRORS.MISSING_PARAMETERS.STATUS_CODE,
-        ERRORS.MISSING_PARAMETERS.MESSAGE,
-      );
-    }
-
-    validateObjectId(author);
-
-    const post = await Post.create({
-      title,
-      author,
-      content,
-    });
-
-    return NextResponse.json({ status: 'success', data: post });
-  } catch (error) {
-    return sendErrorResponse(error);
-  }
-}
-
-export { GET, POST };
+export { GET };
